Changes to video

+++ app/Http/Controllers/PasswordResetController.php
@@ -8,6 +8,7 @@ use App\Notifications\PasswordResetSuccess;
 use App\User;
 use App\PasswordReset;
 use Illuminate\Support\Str;
+use Validator;
 class PasswordResetController extends Controller
 {
     /**
@@ -78,28 +79,35 @@ class PasswordResetController extends Controller
      */
     public function reset(Request $request)
     {
-        $request->validate([
+        $data = Validator::make($request->all(), [
             'email' => 'required|string|email',
-            'password' => 'required|string|confirmed',
+            'password' => 'required|string|confirmed|min:8',
             'token' => 'required|string'
         ]);
-        $passwordReset = PasswordReset::where([
-            ['token', $request->token],
-            ['email', $request->email]
-        ])->first();
-        if (!$passwordReset)
-            return response()->json([
-                'message' => 'This password reset token is invalid.'
-            ], 404);
-        $user = User::where('email', $passwordReset->email)->first();
-        if (!$user)
-            return response()->json([
-                'message' => "We can't find a user with that e-mail address."
-            ], 404);
-        $user->password = bcrypt($request->password);
-        $user->save();
-        $passwordReset->delete();
-        $user->notify(new PasswordResetSuccess($passwordReset));
-        return response()->json($user);
+        if ($data->fails())
+        {
+            return response()->json(['errors' => $data->errors()->all()]);
+        }
+        else {
+            $passwordReset = PasswordReset::where([
+                ['token', $request->token],
+                ['email', $request->email]
+            ])->first();
+            if (!$passwordReset)
+                return response()->json([
+                    'message' => 'This password reset token is invalid.'
+                ]);
+            $user = User::where('email', $passwordReset->email)->first();
+            if (!$user)
+                return response()->json([
+                    'message' => "We can't find a user with that e-mail address."
+                ], 404);
+            $user->password = bcrypt($request->password);
+            $user->save();
+            $passwordReset->delete();
+            $user->notify(new PasswordResetSuccess($passwordReset));
+            return response()->json($user);
+        }
+       
     }
 }
